+++ novelservice/src/main/java/com/iucyh/novelservice/novel/service/registry/NovelPagingStrategyRegistry.java
@@ -1,4 +1,4 @@
-package com.iucyh.novelservice.novel.service.factory;
+package com.iucyh.novelservice.novel.service.registry;
 
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
 import com.iucyh.novelservice.novel.repository.query.paging.NovelPagingStrategy;
@@ -11,15 +11,15 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 /**
- * <p>Novel의 페이징 전략을 조회하는 Factory</p>
+ * <p>Novel의 페이징 전략을 조회하는 Registry</p>
  * <b>빈 생성 시 특정 NovelSortType과 매칭되는 전략이 없으면 {@code IllegalStateException} 발생</b>
  */
 @Component
-public class NovelPagingStrategyFactory {
+public class NovelPagingStrategyRegistry {
 
     private final Map<NovelSortType, NovelPagingStrategy> pagingStrategyMap;
 
-    public NovelPagingStrategyFactory(List<NovelPagingStrategy> pagingStrategies) {
+    public NovelPagingStrategyRegistry(List<NovelPagingStrategy> pagingStrategies) {
         this.pagingStrategyMap = pagingStrategies
                 .stream()
                 .collect(
@@ -42,7 +42,7 @@ public class NovelPagingStrategyFactory {
     /**
      * <p>{@code sortType}과 매칭되는 페이징 전략 조회</p>
      * <b>빈 초기화 단계에서 각 {@code NovelSortType}과 매칭되는 전략이 있는지 검증하므로 null은 절대 반환되지 않음</b>
-     * @param sortType 조회 기준
+     * @param sortType 조회 시 사용할 정렬 기준
      * @return 전달된 {@code sortType}과 매칭되는 {@code NovelPagingStrategy}
      */
     public NovelPagingStrategy get(NovelSortType sortType) {
