+++ novelservice/src/main/java/com/iucyh/novelservice/section/service/SectionQueryService.java
@@ -6,7 +6,7 @@ import com.iucyh.novelservice.novel.enumtype.NovelSortType;
 import com.iucyh.novelservice.novel.repository.query.NovelQueryRepository;
 import com.iucyh.novelservice.novel.repository.query.condition.NovelPagingCondition;
 import com.iucyh.novelservice.novel.repository.query.paging.NovelPagingStrategy;
-import com.iucyh.novelservice.novel.service.factory.NovelPagingStrategyFactory;
+import com.iucyh.novelservice.novel.service.registry.NovelPagingStrategyRegistry;
 import com.iucyh.novelservice.novel.web.dto.mapper.NovelResponseMapper;
 import com.iucyh.novelservice.novel.web.dto.response.NovelSummaryResponse;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +20,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class SectionQueryService {
 
-    private final NovelPagingStrategyFactory pagingStrategyFactory;
+    private final NovelPagingStrategyRegistry pagingStrategyFactory;
     private final NovelQueryRepository novelQueryRepository;
 
     public List<NovelSummaryResponse> getPopularNovels(NovelCategory category) {
